refactor(ffmpeg): codec-specific HW detection and separate HW profiles file
- DetectHardwareEncoder now accepts a Codec parameter so detection is
  specific to the target codec rather than checking any HW encoder
- Add DetectHardwareDecoder(Codec) for symmetric decode-side detection
- Move HW encoder/decoder name constants, hwProfile type, hwProfiles var,
  and hwEncoderNameForCodec/hwDecoderNameForCodecID helpers to hw_profiles.go
- resolveHWAccel now passes the configured video codec to DetectHardwareEncoder
- Fix slices.Contains lint warning in pixel format selection
- Update tests for new DetectHardwareEncoder(Codec) signature
