+++ pkg/ffmpeg/transcode_hw_test.go
@@ -12,13 +12,19 @@ import (
 )
 
 // TestDetectHardwareEncoder_HardwarePresent verifies that DetectHardwareEncoder
-// returns a non-None value when hardware is available. This test only runs when
-// the hwtest build tag is set (i.e. when make test detects hardware via
-// ffmpeg -encoders). Detecting no hardware with the hwtest tag active is a
-// likely bug — the test fails rather than skips in that case.
+// returns a non-None value for at least one supported codec when hardware is
+// available. This test only runs when the hwtest build tag is set (i.e. when
+// make test detects hardware via ffmpeg -encoders). Detecting no hardware with
+// the hwtest tag active is a likely bug — the test fails rather than skips.
 func TestDetectHardwareEncoder_HardwarePresent(t *testing.T) {
-	hw, err := ffmpeg.DetectHardwareEncoder()
-	require.NoError(t, err)
-	assert.NotEqual(t, ffmpeg.HWAccelNone, hw,
-		"DetectHardwareEncoder must return a non-None value when hardware is present")
+	var foundHW bool
+	for _, codec := range []ffmpeg.Codec{ffmpeg.CodecH264, ffmpeg.CodecH265} {
+		hw, err := ffmpeg.DetectHardwareEncoder(codec)
+		require.NoError(t, err)
+		if hw != ffmpeg.HWAccelNone {
+			foundHW = true
+			break
+		}
+	}
+	assert.True(t, foundHW, "DetectHardwareEncoder must return a non-None value for at least one codec when hardware is present")
 }
